style: adjust to designs, small styles refactor
