+++ src/components/gridSettingsDropdownButton/GridSettingsDropdownButton.tsx
@@ -1,15 +1,17 @@
 import { ArrowDropDown } from "@material-ui/icons";
+import classNames from "classnames";
 import React, { FC } from "react";
 import { Button, ButtonProps } from "reactstrap";
+import "./GridSettingsDropdownButton.scss";
 
 export const GridSettingsDropdownButton: FC<ButtonProps> = props => {
-  const { children } = props;
+  const { children, className } = props;
   return (
     <Button
+      {...props}
       outline
       size="sm"
-      className="my-2 btn border-0 d-flex align-items-center text-dark bg-transparent shadow-none"
-      {...props}
+      className={classNames(className, "grid-settings-dropdown-btn")}
     >
       {children}
       <ArrowDropDown className="pl-1" />
