+++ src/components/table/customisations/PagingPanelContent.tsx
@@ -3,6 +3,7 @@ import React, { FC } from "react";
 import { connect } from "react-redux";
 import { labels } from "../../../config/i18n/labels";
 import { DataGridState } from "../DataGridModels";
+import "./PagingPanelContent.scss";
 
 interface Props extends PagingPanel.ContainerProps {
   selectedCount: number;
@@ -10,7 +11,7 @@ interface Props extends PagingPanel.ContainerProps {
 }
 
 export const PagingPanelContent: FC<Props> = ({ selectedCount, ...p }) => (
-  <div className="d-flex align-items-center justify-content-end">
+  <div className="paging-panel-content">
     <div className="mr-n3">{labels.show}</div>
     <PagingPanel.Container {...p} />
   </div>
