+++ src/components/table/customisations/toolbar/ToolbarComponent.scss
@@ -1,0 +1,5 @@
+@import "../../../../styles/variables";
+
+.grid-toolbar {
+  font-size: $grid-toolbar-font-size;
+}
