+++ src/components/table/customisations/toolbar/ToolbarComponent.tsx
@@ -2,12 +2,13 @@ import { Toolbar } from "@devexpress/dx-react-grid-bootstrap4";
 import React, { FC } from "react";
 import { connect, DispatchProp } from "react-redux";
 import { Button } from "reactstrap";
-import { labels } from "../../../config/i18n/labels";
-import { BreakOutOfSubspace } from "../../../utils/subspace/SubspaceProviderHacked";
-import { setFilters } from "../dataGridActions";
-import { DataGridState } from "../DataGridModels";
-import { GridSelectionInfoConnected } from "./GridSelectionInfo";
+import { labels } from "../../../../config/i18n/labels";
+import { BreakOutOfSubspace } from "../../../../utils/subspace/SubspaceProviderHacked";
+import { setFilters } from "../../dataGridActions";
+import { DataGridState } from "../../DataGridModels";
+import { GridSelectionInfoConnected } from "../GridSelectionInfo";
 import { LangSelectorConnected } from "./LangSelector";
+import "./ToolbarComponent.scss";
 import { ViewModeSelectorConnected } from "./ViewModeSelector";
 
 interface Props extends Toolbar.RootProps, DispatchProp {
@@ -21,7 +22,7 @@ export const ToolbarComponent: FC<Props> = ({
 }) => (
   <div
     {...p}
-    className="d-flex flex-wrap justify-content-end align-items-center my-2"
+    className="grid-toolbar d-flex flex-wrap justify-content-end align-items-center my-2"
   >
     {enabledFilters.length > 0 && (
       <Button size="sm" outline onClick={() => dispatch(setFilters([]))}>
